Organize factories and loaders into dedicated folders
