+++ src/engine/loaders/priceLoader.js
@@ -1,4 +1,4 @@
-// src/engine/priceLoader.js
+// src/engine/loaders/priceLoader.js
 import fs from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
@@ -7,7 +7,7 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 // Annahme: <root>/data/config/{devicePrices.json,strainPrices.json}
-const CONFIG_DIR = path.resolve(__dirname, '../../data/config');
+const CONFIG_DIR = path.resolve(__dirname, '../../../data/config');
 
 async function readJson(fp) {
   const raw = await fs.readFile(fp, 'utf-8');
