+++ src/engine/loaders/strainLoader.js
@@ -1,4 +1,4 @@
-// src/engine/strainLoader.js
+// src/engine/loaders/strainLoader.js
 import fs from 'fs/promises';
 import path from 'path';
 import { fileURLToPath } from 'url';
@@ -7,7 +7,7 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 // Assumption: Project structure: <root>/src/engine/* and <root>/data/strains/*.json
-const STRAINS_DIR = path.resolve(__dirname, '../../data/strains');
+const STRAINS_DIR = path.resolve(__dirname, '../../../data/strains');
 
 async function readJson(fp) {
   const raw = await fs.readFile(fp, 'utf-8');
