+++ src/server/services/savegameLoader.js
@@ -5,7 +5,7 @@ import { fileURLToPath } from 'url';
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
-const savegamesDir = path.join(__dirname, '..', '..', 'data', 'savegames');
+const savegamesDir = path.join(__dirname, '..', '..', '..', 'data', 'savegames');
 
 /**
  * Loads and parses a savegame file.
